refactor: 💡 Added refactored the protocol client example
