--- OLD_CODEBASE/examples/react-libs/package.json
@@ -1,59 +1,0 @@
-{
-  "name": "@windingtree/examples-react-libs",
-  "private": true,
-  "version": "0.0.0",
-  "type": "module",
-  "files": [
-    "dist"
-  ],
-  "main": "./dist/index.es.js",
-  "module": "./dist/index.es.js",
-  "types": "./dist/index.d.ts",
-  "exports": {
-    ".": {
-      "import": "./dist/index.es.js"
-    },
-    "./providers": {
-      "import": "./dist/providers.es.js"
-    },
-    "./utils": {
-      "import": "./dist/utils.es.js"
-    }
-  },
-  "devDependencies": {
-    "@types/react": "^18.2.14",
-    "@types/react-dom": "^18.2.6",
-    "@vitejs/plugin-react": "^4.0.0",
-    "eslint": "^8.44.0",
-    "eslint-config-react-app": "^7.0.1",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "typescript": "^5.1.6",
-    "vite": "^4.3.9",
-    "vite-plugin-dts": "^3.0.3"
-  },
-  "peerDependencies": {
-    "viem": "^1.2.9"
-  },
-  "eslintConfig": {
-    "extends": [
-      "react-app"
-    ]
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
-  },
-  "scripts": {
-    "build": "tsc && vite build",
-    "lint": "eslint . --ext .ts"
-  }
-}
