--- OLD_CODEBASE/examples/react-libs/src/index.ts
@@ -1,2 +1,0 @@
-export * as providers from './providers';
-export * as utils from './utils';
