--- OLD_CODEBASE/examples/react-libs/src/providers/index.ts
@@ -1,1 +1,0 @@
-export * as walletProvider from './WalletProvider';
