+++ examples/client/index.html
