+++ examples/client/package.json
@@ -4,17 +4,26 @@
   "version": "0.0.0",
   "type": "module",
   "devDependencies": {
-    "@types/react": "^18.2.14",
-    "@types/react-dom": "^18.2.6",
-    "@vitejs/plugin-react": "^4.0.0",
-    "eslint": "^8.44.0",
+    "@types/react": "^18.2.15",
+    "@types/react-dom": "^18.2.7",
+    "@types/luxon": "^3.3.0",
+    "@vitejs/plugin-react": "^4.0.3",
+    "@windingtree/sdk-react": "workspace:*",
+    "@windingtree/sdk-client": "workspace:*",
+    "@windingtree/sdk-types": "workspace:*",
+    "@windingtree/sdk-storage": "workspace:*",
+    "@windingtree/sdk-messages": "workspace:*",
+    "@windingtree/sdk-utils": "workspace:*",
+    "wtmp-protocol-examples-shared-files": "workspace:*",
+    "eslint": "^8.45.0",
     "eslint-config-react-app": "^7.0.1",
-    "eslint-plugin-react-refresh": "^0.4.1",
+    "eslint-plugin-react-refresh": "^0.4.3",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "typescript": "^5.1.6",
-    "viem": "^1.2.9",
-    "vite": "^4.3.9"
+    "viem": "^1.3.0",
+    "@libp2p/interfaces": "^3.3.2",
+    "luxon": "^3.3.0"
   },
   "eslintConfig": {
     "extends": [
@@ -40,9 +49,8 @@
     ]
   },
   "scripts": {
-    "dev": "vite --force --port 5173",
+    "start": "vite --force --port 5173",
     "build": "tsc && vite build",
-    "preview": "vite preview",
     "lint": "eslint . --ext .ts"
   }
 }
