+++ examples/client/src/components/Requests.tsx
@@ -1,8 +1,8 @@
-import { utils } from '../../../../src/index.js'; // @windingtree/sdk
-import { ClientRequestRecord } from '../../../../src/client/requestsManager.js';
-import { OfferData } from '../../../../src/shared/types.js';
-import { RequestQuery, OfferOptions } from '../../../shared/index.js';
-import { centerEllipsis } from '../../../react-libs/src/utils/index.js';
+import { isExpired } from '@windingtree/sdk-utils';
+import { ClientRequestRecord } from '@windingtree/sdk-client';
+import { OfferData } from '@windingtree/sdk-types';
+import { RequestQuery, OfferOptions } from 'wtmp-protocol-examples-shared-files';
+import { centerEllipsis } from '@windingtree/sdk-react/utils';
 
 export type RequestsRegistryRecord = Required<
   ClientRequestRecord<RequestQuery, OfferOptions>
@@ -51,7 +51,7 @@ export const Requests = ({
               <td>{centerEllipsis(r.data.id)}</td>
               <td>{JSON.stringify(r.data.query)}</td>
               <td>{subscribed && subscribed(r.data.id) ? '✅' : 'no'}</td>
-              <td>{utils.isExpired(r.data.expire) ? '✅' : 'no'}</td>
+              <td>{isExpired(r.data.expire) ? '✅' : 'no'}</td>
               <td>
                 {r.offers.length === 0 ? 0 : ''}
                 {r.offers.length > 0 && (
@@ -61,7 +61,7 @@ export const Requests = ({
                 )}
               </td>
               <td>
-                {!r.subscribed && !utils.isExpired(r.data.expire) && (
+                {!r.subscribed && !isExpired(r.data.expire) && (
                   <button
                     onClick={() => {
                       onCancel(r.data.id);
