+++ examples/client/src/components/Tabs.tsx
