+++ examples/client/src/vite-env.d.ts
