+++ examples/client/vite.config.ts
