+++ package.json
@@ -54,6 +54,7 @@
     "test": "vitest run",
     "coverage": "vitest run --coverage",
     "example:server": "DEBUG=* pnpm -C examples/server run start",
-    "example:node": "DEBUG=* pnpm -C examples/node run start"
+    "example:node": "DEBUG=* pnpm -C examples/node run start",
+    "example:client": "pnpm -C examples/client run start"
   }
 }
